+++ ThirdPartyAdapters/applovin/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Next Version
 - Removed support for Native ads and Medium Rectangle banner ads.
+- Adapter can now load and request ads using an `Application` context.
 
 #### Version 9.13.4.0
 - Verified compatibility with AppLovin SDK 9.13.4.